+++ core/app/c/(public)/[communitySlug]/public/forms/[formSlug]/fill/ExternalFormWrapper.tsx
@@ -7,7 +7,10 @@ import { usePathname, useRouter, useSearchParams } from "next/navigation";
 import type { PubEditorClientProps } from "~/app/components/pubs/PubEditor/PubEditorClient";
 import type { Form as PubPubForm } from "~/lib/server/form";
 import { PubEditorClient } from "~/app/components/pubs/PubEditor/PubEditorClient";
-import { SAVE_STATUS_QUERY_PARAM, SUBMIT_ID_QUERY_PARAM } from "./constants";
+import {
+	SAVE_STATUS_QUERY_PARAM,
+	SUBMIT_ID_QUERY_PARAM,
+} from "~/app/components/pubs/PubEditor/SaveStatus";
 
 const isComplete = (formElements: PubPubForm["elements"], values: FieldValues) => {
 	const requiredElements = formElements.filter((fe) => fe.required && fe.slug);
