--- core/app/c/(public)/[communitySlug]/public/forms/[formSlug]/fill/constants.ts
@@ -1,2 +1,0 @@
-export const SAVE_STATUS_QUERY_PARAM = "saveStatus" as const;
-export const SUBMIT_ID_QUERY_PARAM = "submitId" as const;
