+++ core/app/c/(public)/[communitySlug]/public/forms/[formSlug]/fill/page.tsx
@@ -16,6 +16,7 @@ import { ContextEditorContextProvider } from "~/app/components/ContextEditor/Con
 import { isButtonElement } from "~/app/components/FormBuilder/types";
 import { FormElement } from "~/app/components/forms/FormElement";
 import { FormElementToggleProvider } from "~/app/components/forms/FormElementToggleContext";
+import { SaveStatus, SUBMIT_ID_QUERY_PARAM } from "~/app/components/pubs/PubEditor/SaveStatus";
 import { getLoginData } from "~/lib/authentication/loginData";
 import { getCommunityRole } from "~/lib/authentication/roles";
 import { getPub, getPubCached, getPubs, getPubTypesForCommunity } from "~/lib/server";
@@ -26,10 +27,8 @@ import {
 	renderMarkdownWithPub,
 } from "~/lib/server/render/pub/renderMarkdownWithPub";
 import { capitalize } from "~/lib/string";
-import { SUBMIT_ID_QUERY_PARAM } from "./constants";
 import { ExternalFormWrapper } from "./ExternalFormWrapper";
 import { RequestLink } from "./RequestLink";
-import { SaveStatus } from "./SaveStatus";
 import { handleFormToken } from "./utils";
 
 const NotFound = ({ children }: { children: ReactNode }) => {
